// src/common/enums/prompt.enum.ts
export enum PrimaryObjective {
  CustomerSupport = 'Customer support',
  TechnicalSupport = 'Technical support',
  ProductInquiries = 'Product inquiries',
  SalesSupport = 'Sales support',
  Other = 'Other',
}

export enum Languages {
  English = 'English',
  Vietnamese = 'Vietnamese',
  Spanish = 'Spanish',
  French = 'French',
  Chinese = 'Chinese',
  Other = 'Other',
}

export enum Tone {
  Professional = 'Professional',
  Friendly = 'Friendly',
  Humorous = 'Humorous',
  Serious = 'Serious',
  Casual = 'Casual',
  Other = 'Other',
}

export enum Features {
  FAQ = 'Automatically answer frequently asked questions',
  FindProducts = 'Help find products/services',
  ScheduleAppointments = 'Schedule appointments/transactions',
  CollectInformation = 'Collect customer information',
  ProvideTechnicalSupport = 'Provide technical support',
  ProcessOrders = 'Process orders',
  TrackShipments = 'Track shipments',
  Other = 'Other',
}

export enum Integration {
  CRM = 'CRM',
  OrderManagementSystem = 'Order management system',
  CompanyWebsite = 'Company website',
  SocialMedia = 'Social media',
  EmailMarketingSystem = 'Email marketing system',
  PaymentGateway = 'Payment gateway',
  CustomerSupportSystem = 'Customer support system',
  AnalyticsTool = 'Analytics tool',
  Other = 'Other',
}